+++ src/overture_airflow_provider/render.py
@@ -212,6 +212,7 @@ def _build_render_setup_info(
         "databricks_gpu": databricks_config.gpu,
         "glue_execution_class": glue_config.execution_class,
         "glue_verbose": glue_config.verbose,
+        "glue_output_log_group": glue_config.output_log_group,
         "iam_role_name": glue_config.iam_role_name,
         "codeartifact_domain_owner": package_registry.domain_owner,
         "codeartifact_domain": package_registry.domain,